reorganized events file
