+++ markket/strapi.ts
@@ -8,8 +8,11 @@ const MARKKET_API = process.env.NEXT_PUBLIC_MARKKET_API || 'https://api.markket.
  * @returns
  */
 export const createRecord = async <T>(endpoint: string, data: T) => {
+
+  const _url = new URL(`api/${endpoint}`, MARKKET_API);
+
   try {
-    const response = await fetch(`${MARKKET_API}/api/${endpoint}`, {
+    const response = await fetch(_url.toString(), {
       method: "POST",
       headers: { "Content-Type": "application/json" },
       body: JSON.stringify({ data }),
